Tweaked the unprojection calculation for subtitles
Tweaked the use an angle calculation against the camera's FOV to determine if something is "behind" the camera. This seems to work a lot better for clamping and reduced edge cases of wonkiness. This new calculation does clamp to the bottom corners but it is more obvious which side behind the view the sound is coming from. Ideally in the future these would stack so multiple sounds coming from different locations would be readable